clean_architecture/requirements.txt
Django>=4.2
djangorestframework>=3.14
dependency-injector>=4.41
